finished renaming and moving modules
