+++ src/logic.rs
@@ -1,1 +1,12 @@
+use crate::runtime::LibrettoRuntime;
+use self::lson::Lson;
+
 pub mod lson;
+
+//==================================================================================================
+//          Evaluator
+//==================================================================================================
+
+pub trait LibrettoEvaluator {
+    fn evaluate(&self, runtime: &mut LibrettoRuntime) -> Lson;
+}
